+++ repos/paella-zoom-plugin/changelog.md
@@ -2,4 +2,4 @@
 
 - 2.2.4:
     * Add aria-label to es.upv.paella.canvasZoomInButtonPlugin
-    
+    * Add aria-label to es.upv.paella.canvasZoomOutButtonPlugin